fix(server): move ApiImage import
