+++ packages/server/src/shared/api/decorators/index.ts
@@ -1,4 +1,5 @@
 export * from './ApiControllerDecorator';
+export * from './ApiImageDecorator';
 export * from './ApiObjectIdDecorators';
 export * from './ApiOperationDecorators';
 export * from './PaginationDecorator';
